move geometry_tools to consai_tools
